chore: move PocketIC HTTP gateway tests into a separate target
